// generated/models/Office365UsersModel.ts
export interface GraphUser_V1 {
  /** About Me */
  aboutMe?: string;
  /** true if the account is enabled; otherwise, false. This property is required when a user is created. */
  accountEnabled?: boolean;
  /** Birthday */
  birthday?: string;
  /** Business Phones */
  businessPhones?: string[];
  /** The city in which the user is located. */
  city?: string;
  /** The name of the company in which the user works. */
  companyName?: string;
  /** The country/region in which the user is located; for example, "US" or "UK". */
  country?: string;
  /** The name of the department in which the user works. */
  department?: string;
  /** Display Name */
  displayName?: string;
  /** Given Name */
  givenName?: string;
  /** Hire Date */
  hireDate?: string;
  /** Id */
  id?: string;
  /** Interests */
  interests?: string[];
  /** Job Title */
  jobTitle?: string;
  /** Mail */
  mail?: string;
  /** The mail alias for the user. This property must be specified when a user is created. */
  mailNickname?: string;
  /** Mobile Phone */
  mobilePhone?: string;
  /** My Site */
  mySite?: string;
  /** Office Location */
  officeLocation?: string;
  /** Past Projects */
  pastProjects?: string[];
  /** The postal code for the user's postal address. The postal code is specific to the user's country/region. In the United States of America, this attribute contains the ZIP code. */
  postalCode?: string;
  /** Preferred Language */
  preferredLanguage?: string;
  /** Preferred Name */
  preferredName?: string;
  /** Responsibilities */
  responsibilities?: string[];
  /** Schools */
  schools?: string[];
  /** Skills */
  skills?: string[];
  /** State */
  state?: string;
  /** Street Address */
  streetAddress?: string;
  /** Surname */
  surname?: string;
  /** User Principal Name */
  userPrincipalName?: string;
  /** User Type */
  userType?: string;
}